feat/Dynamic Header and Footer
