+++ src/js/main.js
@@ -1,10 +1,12 @@
 import ProductData from "./ProductData.mjs";
 import ProductList from "./ProductList.mjs";
-import { renderCartLength } from "./utils.mjs";
+import { renderCartLength, loadHeaderFooter } from "./utils.mjs";
+
 
 const dataSource = new ProductData("tents");
 const element = document.querySelector(".product-list");
 const listing = new ProductList("Tents", dataSource, element);
 listing.init();
 
-renderCartLength();
+loadHeaderFooter();
+// renderCartLength();
