+++ src/js/product.js
@@ -1,9 +1,10 @@
 import ProductData from "./ProductData.mjs";
 import ProductDetails from "./ProductDetails.mjs";
-import { getParams } from "./utils.mjs";
+import { getParams, loadHeaderFooter } from "./utils.mjs";
 
 const productId = getParams("product");
 const dataSource = new ProductData("tents");
 
 const product = new ProductDetails(productId, dataSource);
 product.init();
+loadHeaderFooter()
